Fixed #25850 -- Ignored soft applied migrations in consistency check.
Ignored initial migrations that have been soft-applied and may be faked
with the --fake-initial flag in the migration history consistency
check. Does not ignore the initial migration if a later migration in
the same app has been recorded as applied.

Included soft-applied migrations in the pre-migrate project state if
any of its children has been applied.

Thanks to Raphaël Hertzog for the initial patch.